Refactor parallel-letter-frequency usage of threads
The initial implementation of parallel-letter-frequency was unnecessarily complex and poorly efficient. Replaced thread logic with a simpler approach. Originally, the code was creating vec of strings and chunking them before sending to worker threads. Instead, new version joins the input into single string and divides the chars across threads. This not only simplifies the code but potentially improves the performance by eliminating the use of vec. This refactor also removes the