Implement ReadStats and WriteStats struct constructors
In the previous version of the code, the constructors for ReadStats and WriteStats structs were unimplemented. This update implements the constructors for both these structs, taking 'wrapped' as an argument and initializing the 'wrapped' field of the struct with the given argument. This change allows the structs to be properly constructed and the 'wrapped' value to be utilized in further code implementations.